+++ src/Block/LayeredNavigation/RenderLayered/DefaultRenderer.php
@@ -12,6 +12,7 @@ use Emico\Tweakwise\Model\Catalog\Layer\Filter;
 use Emico\Tweakwise\Model\Catalog\Layer\Filter\Item;
 use Emico\Tweakwise\Model\Client\Type\FacetType\SettingsType;
 use Emico\Tweakwise\Model\Config;
+use Emico\Tweakwise\Model\NavigationConfig;
 use Emico\Tweakwise\Model\Seo\FilterHelper;
 use Magento\Framework\View\Element\Template;
 use Magento\Framework\Serialize\Serializer\Json;
@@ -41,20 +42,33 @@ class DefaultRenderer extends Template
     protected $config;
 
     /**
+     * @var NavigationConfig
+     */
+    protected $navigationConfig;
+
+    /**
      * Constructor
      *
      * @param Template\Context $context
      * @param Config $config
+     * @param NavigationConfig $navigationConfig
      * @param FilterHelper $filterHelper
      * @param Json $jsonSerializer
      * @param array $data
      */
-    public function __construct(Template\Context $context, Config $config, FilterHelper $filterHelper, Json $jsonSerializer, array $data = [])
-    {
+    public function __construct(
+        Template\Context $context,
+        Config $config,
+        NavigationConfig $navigationConfig,
+        FilterHelper $filterHelper,
+        Json $jsonSerializer,
+        array $data = []
+    ) {
         parent::__construct($context, $data);
         $this->config = $config;
         $this->filterHelper = $filterHelper;
         $this->jsonSerializer = $jsonSerializer;
+        $this->navigationConfig = $navigationConfig;
     }
 
     /**
@@ -91,19 +105,26 @@ class DefaultRenderer extends Template
     }
 
     /**
+     * @return string
+     */
+    public function getJsSortConfig()
+    {
+        return $this->navigationConfig->getJsSortConfig($this->hasAlternateSortOrder());
+    }
+
+    /**
      * @return boolean
      */
     public function hasAlternateSortOrder()
     {
-        $filter = function (Item $item)
-        {
+        $filter = function (Item $item) {
             return $item->getAlternateSortOrder() !== null;
         };
 
         $items = $this->getItems();
-        $itemsWIthAlternateSortOrder = array_filter($items, $filter);
+        $itemsWithAlternateSortOrder = array_filter($items, $filter);
 
-        return \count($items) === \count($itemsWIthAlternateSortOrder);
+        return \count($items) === \count($itemsWithAlternateSortOrder);
     }
 
     /**
@@ -166,23 +187,6 @@ class DefaultRenderer extends Template
     }
 
     /**
-     * @return string
-     */
-    public function getCssId()
-    {
-        return spl_object_hash($this);
-    }
-
-    /**
-     * @param Item $item
-     * @return string
-     */
-    public function getCssItemId(Item $item)
-    {
-        return spl_object_hash($item);
-    }
-
-    /**
      * @return bool
      */
     public function showCheckbox()
@@ -204,23 +208,6 @@ class DefaultRenderer extends Template
     public function getItemPostfix()
     {
         return $this->escapeHtml($this->getFacetSettings()->getPostfix());
-    }
-
-    /**
-     * @return string
-     */
-    public function getJsNavigationConfig(): string
-    {
-        $navigationOptions = ['hasAlternateSort' => $this->hasAlternateSortOrder()];
-        return $this->config->getJsNavigationConfig($navigationOptions);
-    }
-
-    /**
-     * @return Config
-     */
-    public function getJsUseFormFilters()
-    {
-        return $this->config->getJsUseFormFilters();
     }
 
     /**