+++ src/Block/LayeredNavigation/RenderLayered/SwatchRenderer.php
@@ -8,18 +8,20 @@
 
 namespace Emico\Tweakwise\Block\LayeredNavigation\RenderLayered;
 
+use Emico\Tweakwise\Model\Catalog\Layer\Filter;
 use Emico\Tweakwise\Model\Catalog\Layer\Filter\Item;
+use Emico\Tweakwise\Model\Seo\FilterHelper;
+use Emico\Tweakwise\Model\Swatches\SwatchAttributeResolver;
 use Magento\Catalog\Model\Product;
+use Magento\Catalog\Model\ResourceModel\Eav\AttributeFactory as EavAttributeFactory;
 use Magento\Catalog\Model\ResourceModel\Layer\Filter\AttributeFactory;
+use Magento\Eav\Api\Data\AttributeInterface;
 use Magento\Eav\Model\Entity\Attribute;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\View\Element\Template\Context;
 use Magento\Swatches\Block\LayeredNavigation\RenderLayered;
-use Emico\Tweakwise\Model\Catalog\Layer\Filter;
 use Magento\Swatches\Helper\Data;
 use Magento\Swatches\Helper\Media;
-use Emico\Tweakwise\Model\Config;
-use Emico\Tweakwise\Model\Seo\FilterHelper;
-use Magento\Catalog\Model\ResourceModel\Eav\AttributeFactory as EavAttributeFactory;
 
 class SwatchRenderer extends RenderLayered
 {
@@ -33,11 +35,6 @@ class SwatchRenderer extends RenderLayered
     protected $_template = 'Emico_Tweakwise::product/layered/swatch.phtml';
 
     /**
-     * @var Config
-     */
-    protected $config;
-
-    /**
      * @var Filter
      */
     protected $filter;
@@ -48,14 +45,20 @@ class SwatchRenderer extends RenderLayered
     protected $eavAttributeFactory;
 
     /**
+     * @var SwatchAttributeResolver
+     */
+    protected $swatchAttributeResolver;
+
+    /**
      * SwatchRenderer constructor.
      * @param Context $context
      * @param Attribute $eavAttribute
      * @param AttributeFactory $layerAttribute
      * @param Data $swatchHelper
      * @param Media $mediaHelper
-     * @param Config $config
-     * @param EavAttributeFactory $attributeFactory
+     * @param EavAttributeFactory $eavAttributeFactory
+     * @param FilterHelper $filterHelper
+     * @param SwatchAttributeResolver $swatchAttributeResolver
      * @param array $data
      */
     public function __construct(
@@ -64,21 +67,27 @@ class SwatchRenderer extends RenderLayered
         AttributeFactory $layerAttribute,
         Data $swatchHelper,
         Media $mediaHelper,
-        Config $config,
         EavAttributeFactory $eavAttributeFactory,
         FilterHelper $filterHelper,
+        SwatchAttributeResolver $swatchAttributeResolver,
         array $data = []
-    )
-    {
-        parent::__construct($context, $eavAttribute, $layerAttribute, $swatchHelper, $mediaHelper, $data);
-        $this->config = $config;
+    ) {
+        parent::__construct(
+            $context,
+            $eavAttribute,
+            $layerAttribute,
+            $swatchHelper,
+            $mediaHelper,
+            $data
+        );
         $this->eavAttributeFactory = $eavAttributeFactory;
         $this->filterHelper = $filterHelper;
+        $this->swatchAttributeResolver = $swatchAttributeResolver;
     }
 
     /**
      * @param Filter $filter
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws LocalizedException
      */
     public function setFilter(Filter $filter)
     {
@@ -94,20 +103,67 @@ class SwatchRenderer extends RenderLayered
     }
 
     /**
-     * @return bool
+     * @return array
      */
-    public function getUseFormFilters()
+    public function getSwatchData()
     {
-        return $this->config->getUseFormFilters();
-    }
+        if (false === $this->eavAttribute instanceof Attribute) {
+            throw new \RuntimeException('Magento_Swatches: RenderLayered: Attribute has not been set.');
+        }
 
-    /**
-     * @param Item $item
-     * @return string
-     */
-    public function getCssItemId(Item $item)
-    {
-        return spl_object_hash($item);
+        /**
+         * When this attribute has an id it is an actual magento attribute. If so we can use the parent method to
+         * get the swatches, otherwise it is a mocked attribute see:
+         * @see \Emico\Tweakwise\Model\Catalog\Layer\FilterList\Tweakwise line 105
+        */
+        if ($this->eavAttribute->getId()) {
+            return parent::getSwatchData();
+        }
+
+        // We have a derived swatch filter.
+        $swatchAttributeData = $this->swatchAttributeResolver->getSwatchData($this->filter);
+        // There was no attribute to be found
+        if (!$swatchAttributeData) {
+            return parent::getSwatchData();
+        }
+
+        /** @var AttributeInterface|Attribute $attribute */
+        $attribute = $swatchAttributeData['attribute'];
+        $this->filter->setAttributeModel($attribute);
+        $optionIds = array_values($swatchAttributeData['options']);
+        $optionLabels = array_keys($swatchAttributeData['options']);
+
+        $filterItems = [];
+        foreach ($this->filter->getItems() as $item) {
+            if (!in_array($item->getLabel(), $optionLabels, false)) {
+                continue;
+            }
+
+            $filterItems[$item->getLabel()] = $item;
+        }
+
+
+        $attributeOptions = [];
+        foreach ($attribute->getOptions() as $option) {
+            if (!in_array($option->getValue(), $optionIds, false)) {
+                continue;
+            }
+
+            $filterItem = $filterItems[$option->getLabel()] ?? null;
+            if (!$filterItem) {
+                continue;
+            }
+
+            $attributeOptions[$option->getValue()] = $this->getOptionViewData($filterItem, $option);
+        }
+
+        return [
+            'attribute_id' => $attribute->getId(),
+            'attribute_code' => $attribute->getAttributeCode(),
+            'attribute_label' => $this->filter->getFacet()->getFacetSettings()->getTitle(),
+            'options' => $attributeOptions,
+            'swatches' => $this->swatchHelper->getSwatchesByOptionsId($optionIds),
+        ];
     }
 
     /**
@@ -117,13 +173,5 @@ class SwatchRenderer extends RenderLayered
     public function getItemForSwatch($id)
     {
         return $this->filter->getItemByOptionId($id);
-    }
-
-    /**
-     * @return string
-     */
-    public function getJsNavigationConfig()
-    {
-        return $this->config->getJsNavigationConfig();
     }
 }