+++ src/Model/AttributeSlug.php
@@ -6,7 +6,6 @@
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-
 namespace Emico\Tweakwise\Model;
 
 use Emico\Tweakwise\Model\ResourceModel\AttributeSlug as ResourceModel;