+++ src/Model/Swatches/Plugin.php
@@ -6,18 +6,24 @@
 
 namespace Emico\Tweakwise\Model\Swatches;
 
-use \Magento\Catalog\Api\Data\ProductInterface as Product;
+use Magento\Catalog\Api\Data\ProductInterface as Product;
+use Magento\Swatches\Helper\Data;
 
 class Plugin
 {
     /**
+     * @param Data $subject
+     * @param callable $proceed
      * @param Product $parentProduct
      * @param array $attributes
-     * @param callable $proceed
      * @return
      */
-    public function aroundLoadVariationByFallback($subject, callable $proceed, Product $parentProduct, array $attributes)
-    {
+    public function aroundLoadVariationByFallback(
+        Data $subject,
+        callable $proceed,
+        Product $parentProduct,
+        array $attributes
+    ) {
         $hasArrayValues = false;
         foreach ($attributes as $attribute) {
             if (\is_array($attribute)) {